requirements.txt
watchdog
fastapi
uvicorn[standard]
typer
rich
python-dotenv
chromadb>=0.4.0
sentence-transformers>=2.2.0
sqlite-utils
langchain
colorama
mcp>=1.0.0
tqdm
openai
# pytest httpx are usually for dev but keep them if needed
pytest
httpx
# google-generativeai (Deprecated)
google-genai
tenacity
